+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/ICayenneQueryAssembler.java
@@ -1,10 +1,14 @@
 package io.agrest.cayenne.processor;
 
 import io.agrest.RelatedResourceEntity;
+import io.agrest.id.AgObjectId;
+import io.agrest.meta.AgEntity;
 import io.agrest.runtime.processor.select.SelectContext;
 import org.apache.cayenne.exp.property.Property;
 import org.apache.cayenne.query.ColumnSelect;
 import org.apache.cayenne.query.ObjectSelect;
+
+import java.util.Collection;
 
 /**
  * @since 3.7
@@ -20,6 +24,11 @@ public interface ICayenneQueryAssembler {
             Iterable<P> parentData);
 
     <T> Property<?>[] queryColumns(RelatedResourceEntity<T> entity);
+
+    /**
+     * @since 5.0
+     */
+    <T> ObjectSelect<T> createQueryForIds(AgEntity<T> entity, Collection<AgObjectId> ids);
 
 }